refactor: import paths for link error
