+++ src/commands/data/create/record.ts
@@ -8,7 +8,7 @@
 import { Messages } from '@salesforce/core';
 import { SaveResult } from 'jsforce';
 import { SfCommand, Flags } from '@salesforce/sf-plugins-core';
-import { orgFlags, perflogFlag } from '../../../../src/flags';
+import { orgFlags, perflogFlag } from '../../../flags';
 import { stringToDictionary, collectErrorMessages } from '../../../dataCommand';
 
 Messages.importMessagesDirectory(__dirname);
