+++ src/commands/data/delete/bulk.ts
@@ -9,7 +9,7 @@ import { ReadStream } from 'fs';
 import { Connection, Messages, SfError } from '@salesforce/core';
 import { SfCommand, Flags, Ux } from '@salesforce/sf-plugins-core';
 import { Duration } from '@salesforce/kit';
-import { orgFlags } from '../../../../src/flags';
+import { orgFlags } from '../../../flags';
 import { Batcher, BatcherReturnType } from '../../../batcher';
 
 Messages.importMessagesDirectory(__dirname);
