+++ src/commands/data/get/record.ts
@@ -9,7 +9,7 @@ import { Messages, SfError } from '@salesforce/core';
 import { Record } from 'jsforce';
 import { toAnyJson } from '@salesforce/ts-types';
 import { SfCommand, Flags } from '@salesforce/sf-plugins-core';
-import { orgFlags, perflogFlag } from '../../../../src/flags';
+import { orgFlags, perflogFlag } from '../../../flags';
 import { query, logNestedObject } from '../../../dataCommand';
 
 Messages.importMessagesDirectory(__dirname);
