+++ src/commands/data/update/record.ts
@@ -8,7 +8,7 @@
 import { Messages, SfError } from '@salesforce/core';
 import { SaveError, SaveResult } from 'jsforce';
 import { SfCommand, Flags } from '@salesforce/sf-plugins-core';
-import { orgFlags } from '../../../../src/flags';
+import { orgFlags } from '../../../flags';
 import { collectErrorMessages, query, stringToDictionary } from '../../../dataCommand';
 
 Messages.importMessagesDirectory(__dirname);
